Create Beautiful_triplet.py
we know that aar[i] is there is in the array we just asked if arr[i]+d and arr[i]+2*d  is there in the array or not
if it is there then we increased the value of the counter. And then printed that on output stream
This code will clear all the 10 test cases.